+++ src/main/kotlin/com/whalesea/ideatabmanager/toolwindow/TabGroupIcons.kt
@@ -18,6 +18,7 @@ object TabGroupIcons {
     val collapse: Icon = IconLoader.getIcon("/icons/collapseGroup.svg", TabGroupIcons::class.java)
     val expand: Icon = IconLoader.getIcon("/icons/expandGroup.svg", TabGroupIcons::class.java)
     val groupDragHandle: Icon = IconLoader.getIcon("/icons/groupDragHandle.svg", TabGroupIcons::class.java)
+    val tabDragHandle: Icon = IconLoader.getIcon("/icons/tabDragHandle.svg", TabGroupIcons::class.java)
     val newEmptyGroup: Icon = IconLoader.getIcon("/icons/newEmptyGroup.svg", TabGroupIcons::class.java)
     val saveCurrentTabs: Icon = IconLoader.getIcon("/icons/saveCurrentTabs.svg", TabGroupIcons::class.java)
     val saveSelectedTabs: Icon = IconLoader.getIcon("/icons/saveSelectedTabs.svg", TabGroupIcons::class.java)